[TASK] delete item and close application confirmation

+++ Stock/Products.cs
@@ -131,24 +131,28 @@ namespace Stock
 
         private void button1_Click(object sender, EventArgs e)
         {
-            if (Validation())
+            DialogResult dialogResult = MessageBox.Show("Are You Sure Want to Delete", "Message", MessageBoxButtons.YesNo);
+            if (dialogResult == DialogResult.Yes)
             {
-                SqlConnection con = Connection.GetConnection();
-                var sqlQuery = "";
-                if (ifProductExist(con, textBox1.Text))
+                if (Validation())
                 {
-                    con.Open();
-                    sqlQuery = @"DELETE FROM [Products] WHERE [ProductCode] = '" + textBox1.Text + "'";
-                    SqlCommand cmd = new SqlCommand(sqlQuery, con);
-                    cmd.ExecuteNonQuery();
-                    con.Close();
+                    SqlConnection con = Connection.GetConnection();
+                    var sqlQuery = "";
+                    if (ifProductExist(con, textBox1.Text))
+                    {
+                        con.Open();
+                        sqlQuery = @"DELETE FROM [Products] WHERE [ProductCode] = '" + textBox1.Text + "'";
+                        SqlCommand cmd = new SqlCommand(sqlQuery, con);
+                        cmd.ExecuteNonQuery();
+                        con.Close();
+                    }
+                    else
+                    {
+                        MessageBox.Show("Record Not Exist...!");
+                    }
+                    LoadData();
+                    ResetRecords();
                 }
-                else
-                {
-                    MessageBox.Show("Record Not Exist...!");
-                }
-                LoadData();
-                ResetRecords();
             }
         }
         private void ResetRecords()
@@ -167,7 +171,22 @@ namespace Stock
         private bool Validation()
         {
             bool result = false;
-            if (!string.IsNullOrEmpty(textBox1.Text) && !string.IsNullOrEmpty(textBox2.Text) && comboBox1.SelectedIndex > -1 )
+            if (string.IsNullOrEmpty(textBox1.Text))
+            {
+                errorProvider1.Clear();
+                errorProvider1.SetError(textBox1, "Product Code Required!");
+            }
+            else if (string.IsNullOrEmpty(textBox2.Text))
+            {
+                errorProvider1.Clear();
+                errorProvider1.SetError(textBox2, "Product Name Required!");
+            } 
+            else if (comboBox1.SelectedIndex ==- -1 )
+            {
+                errorProvider1.Clear();
+                errorProvider1.SetError(comboBox1, "Select Status");
+            }
+            else
             {
                 result = true;
             }
